pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "powerx-strategy"
version = "0.1.0"
description = "PowerX Strategy with Basic Monte Carlo for MT5"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "matplotlib>=3.7.5",
    "MetaTrader5>=5.0.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "PyYAML>=6.0",
    "ta-lib>=0.4.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.5b2",
    "isort>=5.0.0",
    "flake8>=3.9.0",
]

[tool.setuptools]
packages = ["powerx"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.uv.sources]
ta-lib = { path = "ta_lib-0.6.3-cp313-cp313-win_amd64.whl" }
